nostalgiasite/nostalgiasite/wsgi.py
import os
from django.core.wsgi import get_wsgi_application
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'nostalgiaSite.settings')

application = get_wsgi_application()